20.11.2022

+++ Messenger/twitter_app/views.py
@@ -10,19 +10,19 @@ from django.views import View
 from random import randint
 import datetime
 
-
-from twitter_app.models import Tweets
+from twitter_app.models import Tweets, Rating
 from log_auth.models import Users
 
 
 def new_tweet(request: HttpRequest) -> Union[HttpResponseBadRequest, JsonResponse]:
     if request.method == 'POST':
-        id_tweet = (str(datetime.datetime.now()).replace(' ', '').replace(':', '').replace('-', '').replace(
-            '.', ''))[11:] + str(randint(10, 99))
+        # id_tweet = (str(datetime.datetime.now()).replace(' ', '').replace(':', '').replace('-', '').replace(
+        #     '.', ''))[11:] + str(randint(10, 99))
         author_id = request.POST.get('author_id', None)
         tweet_text = request.POST.get('tweet_text', None)
-        author = Users.objects.get(user_id=int(author_id))
-        Tweets.objects.create(id_tweet=id_tweet, author_id=author_id, author_nickname=author.user_nickname,
+        print(author_id)
+        author = Users.objects.get(id=int(author_id))
+        Tweets.objects.create(author_id=author_id, author_nickname=author.user_nickname,
                               text_tweet=tweet_text, likes=0)
         return redirect(reverse('twitter_app:home', args=(author_id,)))
         # return JsonResponse({'status': 'Todo added!'})
@@ -47,13 +47,26 @@ def get_info_tweet(request: HttpRequest):
     return JsonResponse({'status': 'Invalid request'}, status=400)
 
 
-def like_tweet(request: HttpRequest, id_tweet: int):
+def like_tweet(request: HttpRequest, user_id: int, id_tweet: int):
     if request.method == 'GET':
-        post = Tweets.objects.get(id_tweet=id_tweet)
-        post.likes = Tweets.objects.get(id_tweet=id_tweet).likes + 1
+        print("user_id", user_id)
+        print("id_tweet", id_tweet)
+        # print(Rating.objects.get(user_id=user_id))
+        post = Tweets.objects.get(id=id_tweet)
         author_id = post.author_id
-        post.save()
-        return redirect(reverse('twitter_app:home', args=(author_id,)))
+
+        try:
+            print(Rating.objects.get(user_id=user_id, id_tweet=id_tweet))
+            return redirect(reverse('twitter_app:home', args=(author_id,)))
+        except:
+            Rating.objects.create(
+                id_tweet=id_tweet,
+                user_id=user_id,
+            )
+
+            post.likes = Tweets.objects.get(id=id_tweet).likes + 1
+            post.save()
+            return redirect(reverse('twitter_app:home', args=(author_id,)))
     return JsonResponse({'status': 'Invalid request'}, status=400)
     # return render(request, 'django_twitter_app/home.html', context=context)
 
@@ -69,13 +82,17 @@ def dislike_tweet(request: HttpRequest, id_tweet: int):
     # return render(request, 'django_twitter_app/home.html', context=context)
 
 
-def delete_tweet(request: HttpRequest, id_tweet: int):
+def delete_tweet(request: HttpRequest, id_user: int, id_tweet: int):
     context = {}
     if request.method == 'GET':
-        post = Tweets.objects.get(id_tweet=id_tweet)
+        print(id_tweet)
+        post = Tweets.objects.get(id=id_tweet)
         author_id = post.author_id
-        post.delete()
-        return redirect(reverse('twitter_app:home', args=(author_id,)))
+        if id_user == author_id:
+            post.delete()
+            return redirect(reverse('twitter_app:home', args=(author_id,)))
+        else:
+            return redirect(reverse('twitter_app:home', args=(id_user,)))
     return JsonResponse({'status': 'Invalid request'}, status=400)
 
 
@@ -86,7 +103,6 @@ def delete_tweet(request: HttpRequest, id_tweet: int):
 #
 
 
-
 def home(request: HttpRequest, code: int = 0) -> HttpResponse:
     if request.method == "GET":
         user_id = request.COOKIES.get('user_id')
@@ -95,8 +111,16 @@ def home(request: HttpRequest, code: int = 0) -> HttpResponse:
             return redirect(reverse('log_auth:log_auth', args=()))
         else:
             if int(user_id) == int(code):
+                likes_user = Rating.objects.filter(user_id=user_id)
+                mass_likes_user = []
+                for like_user in likes_user:
+                    mass_likes_user.append(like_user.id_tweet)
+                print(mass_likes_user)
                 todos = list(Tweets.objects.all().values())
                 print(todos)
-                return render(request, 'public/twitter_app/home.html', context={'user_id': user_id, 'content': todos})
+                user_nickname = Users.objects.get(id=user_id).user_nickname
+                return render(request, 'public/twitter_app/home.html',
+                              context={'user_id': user_id, 'content': todos, 'user_nickname': user_nickname,
+                                       'mass_likes_user': mass_likes_user})
             else:
                 return redirect(reverse('twitter_app:home', args=(user_id,)))
